fix: only admins can create,edit,delete a genre

+++ src/gender/gender.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { User } from './../user/entities/user.entity';
+import {
+  Injectable,
+  NotFoundException,
+  UnauthorizedException,
+} from '@nestjs/common';
 import { CreateGenderDto } from './dto/create-gender.dto';
 import { UpdateGenderDto } from './dto/update-gender.dto';
 import { Gender } from './entities/gender.entity';
@@ -14,7 +19,13 @@ export class GenderService {
     updatedAt: false,
   };
   constructor(private readonly prisma: PrismaService) {}
-  async create(createGenderDto: CreateGenderDto): Promise<Gender> {
+  async create(user: User, createGenderDto: CreateGenderDto): Promise<Gender> {
+    if (!user.isAdmin) {
+      throw new UnauthorizedException(
+        'apenas administradores podem criar novos generos',
+      );
+    }
+
     const gender = createGenderDto.name.toLocaleUpperCase();
     const genderToCreate: CreateGenderDto = {
       name: gender,
@@ -50,8 +61,14 @@ export class GenderService {
     return record;
   }
 
-  async update(id: string, updateGenderDto: UpdateGenderDto) {
+  async update(user: User, id: string, updateGenderDto: UpdateGenderDto) {
     await this.findOne(id);
+    if (!user.isAdmin) {
+      throw new UnauthorizedException(
+        'apenas administradores podem editar generos',
+      );
+    }
+
     const data: string = updateGenderDto.name.toLocaleUpperCase();
     const dataToUpdate: UpdateGenderDto = {
       name: data,
@@ -66,8 +83,13 @@ export class GenderService {
       .catch(handleError);
   }
 
-  async remove(id: string) {
+  async remove(user: User, id: string) {
     let genderDeleted = await this.findOne(id);
+    if (!user.isAdmin) {
+      throw new UnauthorizedException(
+        'apenas administradores podem deletar generos',
+      );
+    }
     await this.prisma.genres.delete({ where: { id } }).catch(handleError);
     return `Genero ${genderDeleted.name} deletado com sucesso`;
   }
